LayoutTests/model-element/model-element-stagemode.html

<!DOCTYPE html> <!-- webkit-test-runner [ ModelElementEnabled=true ModelProcessEnabled=true ] -->
<meta charset="utf-8">
<title>&lt;model> stagemode attribute</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/model-element-test-utils.js"></script>
<script src="resources/model-utils.js"></script>
<body>
<script>
'use strict';

promise_test(async t => {
    const model = document.createElement("model");
    const firstSource = model.appendChild(makeSource("resources/cube.usdz"));
    await model.ready;

    assert_equals(model.stageMode, "");
    model.stageMode = "orbit";
    assert_equals(model.stageMode, "orbit");

    const secondSource = model.appendChild(makeSource("resources/heart.usdz"));
    firstSource.remove();
    assert_equals(model.currentSrc, secondSource.src);
    await model.ready;

    assert_equals(model.stageMode, "orbit");
}, `Setting stage mode persists after changing source on the same model element`);

promise_test(async t => {
    const [model, source] = createModelAndSource(t, "resources/cube.usdz");
    await model.ready;

    assert_equals(model.stageMode, "");
    model.stageMode = "orbit";
    assert_equals(model.stageMode, "orbit");
    model.stageMode = "none";
    assert_equals(model.stageMode, "none");
    model.stageMode = "Hello world!";
    assert_equals(model.stageMode, "Hello world!");

}, `Checking <model> stage mode value after updating`);

</script>
</body>
